day 1: minor improvements

+++ utility/sort.c
@@ -1,4 +1,5 @@
 #include "sort.h"
+#include <stdlib.h>
 
 void swap(int *array, int a, int b) {
 	int temp = array[a];
@@ -23,7 +24,7 @@ int partition(int *array, int low, int high) {
 
 void quicksort(int *array, int low, int high){
 	if(low < high) {
-		int pivotIndex = partition(array, low, high);
+		int pivotIndex = partition(array, (int)low, high);
 		quicksort(array, low, pivotIndex - 1);
 		quicksort(array, pivotIndex + 1, high);
 	}
